Escape underscores (#208)
* IdentifierConverter: Add the option to not escape underscores

This allows for generating symbols with subscripts (possibly multiple layers), e.g. `x_2_i` becomes `x_{2_{i}}`. This option also creates the possibility of generating invalid latex if the identifier starts or ends with an underscore, or has a double underscore in it. In these cases we now raise a ValueError.

* CodeGen: Add the plumbing to allow the use of the new `escape_underscores` option in IdentifierConverter

* IdentifierCodegen: Ensure all subscripts are wrapped in braces

Previously generated latex was incorrect for multi-character subscripts.
Tests updated to reflect new (correct) behaviour.

* IdentifierConverter: Simplify error checking for invalid LaTeX when escaping underscores

* IdentifierConverter: Fix too-long lines that flake8 complains about